Move service to scripts folder
